End-To-End-ChatBot/requirements.txt
# requirements.txt
streamlit==1.38.0
python-dotenv==1.0.1
requests==2.32.3

prometheus-client==0.20.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0

langchain==0.3.5
langchain-openai==0.1.22
langchain-core==0.3.12
pydantic==2.9.2

guardrails-ai==0.5.0
PyYAML==6.0.2
promptlayer==0.2.12
